% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_management.R
\name{get_swat_parameters}
\alias{get_swat_parameters}
\title{Retrieve SWAT model parameters}
\usage{
get_swat_parameters(version = "SWAT")
}
\arguments{
\item{version}{character. SWAT version ("SWAT" or "SWAT_T"). Default:
"SWAT".}
}
\value{
A tibble with:
\itemize{
  \item `component`: SWAT input component (e.g., ".gw", ".mgt").
  \item `parameter`: Parameter name.
  \item `line`: Line index in the file.
  \item `start`: Start position in the line.
  \item `stop`: End position in the line.
  \item `format`: Print format used when writing.
  \item `description`: Short description.
}
}
\description{
Returns the set of SWAT parameters available for calibration or sensitivity
analysis for a given SWAT version. The table helps define valid
`component`/`parameter` pairs for \link{change_parameter} and
\link{change_multiple_parameters}.
}
\examples{
# Parameters for default SWAT
p <- get_swat_parameters()

# Parameters for SWAT_T
pt <- get_swat_parameters(version = "SWAT_T")

}
\seealso{
Other Parameter management: 
\code{\link{change_multiple_parameters}()},
\code{\link{change_parameter}()},
\code{\link{split_swat_parameters}()}
}
\concept{Parameter management}
