% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gridded_climate_netcdfs}
\alias{gridded_climate_netcdfs}
\title{Gridded climate NetCDFs (precipitation and temperature)}
\format{
Three NetCDF files:
\describe{
  \item{pr.nc}{Daily precipitation (mm/day), 7 layers.}
  \item{tasmax.nc}{Daily maximum temperature (deg C), 7 layers.}
  \item{tasmin.nc}{Daily minimum temperature (deg C), 7 layers.}
}
}
\source{
\itemize{
  \item RAIN4PE: \doi{10.5880/pik.2020.010}
  \item PISCO temperature v1.1:
  \url{https://iridl.ldeo.columbia.edu/SOURCES/.SENAMHI/.HSR/.PISCO/.Temp/.v1p1/}
}
}
\description{
Small NetCDFs with daily precipitation (RAIN4PE) and daily min/max temperature
(PISCO). Period: 7 days in Jan 2010. Resolution: 0.1 deg (WGS84; EPSG:4326).
}
\details{
Files are located under \code{inst/extdata} and should be accessed with
\code{system.file()}.
}
\examples{
pr_nc     <- system.file("extdata", "pr.nc", package = "hydroSWAT")
tasmax_nc <- system.file("extdata", "tasmax.nc", package = "hydroSWAT")
tasmin_nc <- system.file("extdata", "tasmin.nc", package = "hydroSWAT")
# r <- terra::rast(pr_nc)
# terra::plot(r[[1]], main = "Precipitation (mm/day) 2010-01-01")

}
\seealso{
Other Data: 
\code{\link{example_basin_datasets}},
\code{\link{qobserved}},
\code{\link{riv1}},
\code{\link{subs1}},
\code{\link{swat_txtinout_data}}
}
\concept{Data}
\keyword{datasets}
