% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_readers.R
\name{output_hru}
\alias{output_hru}
\title{Read SWAT HRU output}
\usage{
output_hru(
  file = "output.hru",
  variable = NULL,
  target_id = NULL,
  time_step = "daily",
  output_start_date
)
}
\arguments{
\item{file}{Character. Absolute path to the \code{output.hru} file. Default
is "output.hru".}

\item{variable}{Character vector with the names of variables to read. Use
\code{\link{get_swat_vars}("hru")} to list all options. If \code{NULL}, all
variables are read.}

\item{target_id}{Character vector of HRU IDs (HRUGIS). Default \code{NULL}
retrieves all HRUs.}

\item{time_step}{Character. Time step of the simulation ("daily" or
"monthly"). Default is "daily".}

\item{output_start_date}{Character. Date when SWAT output generation begins
(first day after warm-up), in "yyyy-mm-dd" format.}
}
\value{
A \code{data.frame} with the selected variables for each HRU and time
step. Columns "LULC", "HRU", "GIS", "SUB", and "MON" are always included.
}
\description{
Reads the \code{output.hru} file generated by SWAT, allowing users to filter
the data by variables, subbasin IDs, HRU IDs, and time steps (daily or
monthly).
}
\examples{
# Example with package data
tmpdir <- tempdir()
get_swat_example(tmpdir)
hru_file <- file.path(tmpdir, "TxtInOut", "output.hru")
hru_data <- output_hru(file = hru_file, variable = c("PRECIPmm", "ETmm", "WYLDmm"),
                       target_id = NULL, time_step = "daily",
                       output_start_date = "2011-01-01")
head(hru_data)

\dontrun{
# Example with external SWAT project
hru_data <- output_hru(file = "C:/SWAT_Project/TxtInOut/output.hru",
                       variable = NULL, target_id = NULL,
                       time_step = "monthly",
                       output_start_date = "2000-01-01")
}
}
\seealso{
Other Output readers: 
\code{\link{get_swat_vars}()},
\code{\link{output_rch}()},
\code{\link{output_sub}()},
\code{\link{read_pcp}()},
\code{\link{read_tmp}()}
}
\concept{Output readers}
