% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance-04-plots.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series of simulated and observed data}
\usage{
plot_timeseries(
  data,
  x_label = "Time",
  y_label = "Q [m3/s]",
  show_legend = TRUE
)
}
\arguments{
\item{data}{data frame. Time series of simulated and observed values,
typically one of the outputs from \link{evaluate_swat} (daily_data,
monthly_data, or annual_data).}

\item{x_label}{character. Label for the x-axis. Default is "Time".}

\item{y_label}{character. Label for the y-axis. Default is "Q [m3/s]".}

\item{show_legend}{logical. Whether to display the legend. Default is TRUE.}
}
\value{
A ggplot object representing the time series plot of observed and
simulated values.
}
\description{
Generates a time series plot comparing simulated and observed values
(e.g., streamflow) from the outputs of \link{evaluate_swat}.
}
\examples{
tmpdir <- tempdir()
get_swat_example(tmpdir)
rch_file <- file.path(tmpdir, "TxtInOut", "output.rch")
rch_data <- output_rch(
  file = rch_file, variable = "FLOW_OUTcms",
  target_id = 3, time_step = "daily",
  output_start_date = "2011-01-01"
)

observed_data <- tibble::tibble(
  Date = qobserved$Date,
  value = qobserved$Flow,
  target_id = 3
)

evaluation_results <- evaluate_swat(
  rch_output = rch_data,
  observed_data = observed_data,
  target_id = 3,
  variable = "FLOW_OUTcms",
  metrics = c("NSE","KGE","MAE","PBIAS"),
  start_dates = c("2011-01-01","2014-01-01"),
  end_dates = c("2013-12-31","2015-12-31")
)

# Plot daily time series of observed vs simulated flow
plot_timeseries(evaluation_results$daily_data)
}
\seealso{
Other Performance evaluation: 
\code{\link{evaluate_swat}()},
\code{\link{fdc_signature}()},
\code{\link{gof_metrics}()},
\code{\link{pbias_fdc_midsegment}()},
\code{\link{pbias_fdc_target_segment}()},
\code{\link{plot_fdc}()},
\code{\link{plot_mean_annual_cycle}()}
}
\concept{Performance evaluation}
