% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{qobserved}
\alias{qobserved}
\title{Observed streamflow (daily)}
\format{
Tibble/data.frame with:
\describe{
  \item{Date}{Date (daily).}
  \item{Flow}{Discharge (m^3/s).}
}
}
\source{
National Meteorology and Hydrology Service of Peru (SENAMHI).
}
\usage{
qobserved
}
\description{
Daily discharge at the outlet gauge of the example basin.
}
\examples{
head(qobserved)
# plot(qobserved$Date, qobserved$Flow, type = "l",
#      xlab = "Date", ylab = "Flow (m^3/s)")

}
\seealso{
Other Data: 
\code{\link{example_basin_datasets}},
\code{\link{gridded_climate_netcdfs}},
\code{\link{riv1}},
\code{\link{subs1}},
\code{\link{swat_txtinout_data}}
}
\concept{Data}
\keyword{datasets}
