% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{riv1}
\alias{riv1}
\title{River network (sf)}
\format{
An \code{sf} object with line features and related attributes.
}
\source{
Prepared by the hydroSWAT authors.
}
\usage{
riv1
}
\description{
River reaches used by the example SWAT project.
}
\examples{
# plot(riv1["Subbasin"])

}
\seealso{
Other Data: 
\code{\link{example_basin_datasets}},
\code{\link{gridded_climate_netcdfs}},
\code{\link{qobserved}},
\code{\link{subs1}},
\code{\link{swat_txtinout_data}}
}
\concept{Data}
\keyword{datasets}
