% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_management.R
\name{split_swat_parameters}
\alias{split_swat_parameters}
\title{Split SWAT parameters for parallel processing}
\usage{
split_swat_parameters(cores, parameter_sample)
}
\arguments{
\item{cores}{integer. Number of chunks to produce (e.g., CPU cores).}

\item{parameter_sample}{matrix or data.frame. Table of parameter sets
to split (one row per set).}
}
\value{
A list of length `cores` with subsets of `parameter_sample` (input row
order preserved).
}
\description{
Split a table of parameter sets into `cores` contiguous chunks for
parallel execution, distributing rows as evenly as possible.
}
\examples{
\donttest{
x <- data.frame(a = 1:10, b = 11:20)
split <- split_swat_parameters(cores = 3, parameter_sample = x)
vapply(split, nrow, integer(1))
}

}
\seealso{
Other Parameter management: 
\code{\link{change_multiple_parameters}()},
\code{\link{change_parameter}()},
\code{\link{get_swat_parameters}()}
}
\concept{Parameter management}
