% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{subs1}
\alias{subs1}
\title{Subbasins (sf)}
\format{
An \code{sf} object with polygon features and related attributes.
}
\source{
Prepared by the hydroSWAT authors.
}
\usage{
subs1
}
\description{
Subbasins used by the example SWAT project.
}
\examples{
# plot(subs1["Subbasin"])

}
\seealso{
Other Data: 
\code{\link{example_basin_datasets}},
\code{\link{gridded_climate_netcdfs}},
\code{\link{qobserved}},
\code{\link{riv1}},
\code{\link{swat_txtinout_data}}
}
\concept{Data}
\keyword{datasets}
