% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{swat_txtinout_data}
\alias{swat_txtinout_data}
\title{SWAT TxtInOut inputs}
\format{
Named list of character vectors. Each element is the content of one SWAT input
file (e.g., \code{file.cio}, \code{.gw}, \code{pcp*.pcp}).
}
\source{
Prepared by the hydroSWAT authors.
}
\usage{
swat_txtinout_data
}
\description{
Named list that mirrors a SWAT \code{TxtInOut} directory from the example project.
}
\examples{
str(swat_txtinout_data[1:3])

}
\seealso{
Other Data: 
\code{\link{example_basin_datasets}},
\code{\link{gridded_climate_netcdfs}},
\code{\link{qobserved}},
\code{\link{riv1}},
\code{\link{subs1}}
}
\concept{Data}
\keyword{datasets}
