% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_tools.R
\name{tmp2swat_editor}
\alias{tmp2swat_editor}
\title{Generate SWAT Editor-compatible temperature input files}
\usage{
tmp2swat_editor(
  grid_tmax,
  grid_tmin,
  subs1,
  start_date,
  time_step = "day",
  unit = "C",
  cores = 1,
  grid_tmax_file = NULL,
  grid_tmin_file = NULL,
  output_dir = getwd()
)
}
\arguments{
\item{grid_tmax}{A `SpatRaster` object representing maximum temperature.}

\item{grid_tmin}{A `SpatRaster` object representing minimum temperature.}

\item{subs1}{An `sf` object representing the subbasins (e.g., shapefile from
ArcSWAT or QSWAT).}

\item{start_date}{Start date of the record in "yyyy-mm-dd" format.}

\item{time_step}{Time step between raster layers. Either 'day' or 'month'.
Default is 'day'.}

\item{unit}{Temperature unit. Either 'C' (Celsius) or 'K' (Kelvin).
Default is 'C'.}

\item{cores}{Number of CPU cores for parallel processing. Default is 1.}

\item{grid_tmax_file}{Path to the NetCDF file for Tmax (required if cores >1).}

\item{grid_tmin_file}{Path to the NetCDF file for Tmin (required if cores >1).}

\item{output_dir}{Directory where SWAT input files will be written. Defaults
to current working directory.}
}
\value{
Writes SWAT Editor-compatible temperature input files and a station
table in the specified directory.
}
\description{
Writes SWAT Editor-compatible temperature input files for each subbasin
by estimating areal mean maximum and minimum temperature from raster or
NetCDF data using \link{areal_mean}.
}
\examples{
\donttest{
# Example with package NetCDF data
tmax_nc <- system.file("extdata", "tasmax.nc", package = "hydroSWAT")
tmin_nc <- system.file("extdata", "tasmin.nc", package = "hydroSWAT")
data(subs1, package = "hydroSWAT")
grid_tmax <- terra::rast(tmax_nc)
grid_tmin <- terra::rast(tmin_nc)
tmp2swat_editor(grid_tmax, grid_tmin, subs1, "2010-01-01", output_dir = tempdir())
}

\dontrun{
# Example with external NetCDF and shapefile
subs1 <- sf::st_read("path_to_subbasin_shapefile.shp")
grid_tmax <- terra::rast("path_to_tmax_data.nc")
grid_tmin <- terra::rast("path_to_tmin_data.nc")
tmp2swat_editor(grid_tmax, grid_tmin, subs1, "1981-01-01", cores = 3,
                grid_tmax_file = "path_to_tmax_data.nc",
                grid_tmin_file = "path_to_tmin_data.nc",
                output_dir = "C:/swat_inputs")
}
}
\seealso{
\code{\link{areal_mean}}, \code{\link{tmp2swat}}

Other Preprocessing tools: 
\code{\link{areal_mean}()},
\code{\link{pcp2swat}()},
\code{\link{pcp2swat_editor}()},
\code{\link{tmp2swat}()}
}
\concept{Preprocessing tools}
